fix(http): remove view handler
